% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_preproc.R
\name{codebook_recoder}
\alias{codebook_recoder}
\title{Codebook Recoder}
\usage{
codebook_recoder(df, var_regex, values_from, values_to, FUN = as.numeric)
}
\arguments{
\item{df}{Dataframe of data to be recoded.}

\item{var_regex}{Vector of regex patterns describing columns to recode values.}

\item{values_from}{Vector of possible values to be recoded.}

\item{values_to}{Vector of values to recode values_from into.}

\item{FUN}{Function to to apply to column after recoding. Set to NULL if no function is desired.}
}
\value{
A dataframe with recoded values.
}
\description{
Recode the values of multiple columns .
}
