% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_preproc.R
\name{codebook_renamer}
\alias{codebook_renamer}
\title{Codebook Renamer}
\usage{
codebook_renamer(df, names_from, names_to)
}
\arguments{
\item{df}{Dataframe with columns to be renamed}

\item{names_from}{Vector of column names found in df}

\item{names_to}{Vector of column names with the same length as names_from to which columns in df will be renamed}
}
\value{
A dataframe with renamed columns
}
\description{
Rename the columns of a dataframe.
}
