% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_date_dir.R
\name{recent_date_dir}
\alias{recent_date_dir}
\title{recent_date_dir}
\usage{
recent_date_dir(directory, recursive = FALSE)
}
\arguments{
\item{directory}{string: top-level directory in which to search for date directories}

\item{recursive}{boolean: indicates whether to search recurvsively}
}
\value{
string: path of most recent dated directory
}
\description{
recent_date_dir
}
