% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale}
\alias{scale}
\title{Scale}
\usage{
scale(...)
}
\arguments{
\item{...}{Arguments passed to scale}
}
\value{
Vector of scaled values
}
\description{
Wrapper for base scale function. The base scale function returns a list type and this function returns a numeric type.
}
\examples{

myVec <- c(1,6,3,6,8,4,5,4,2,2)
myScaledVec <- scale(myVec)

}
