% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_sd_categories}
\alias{to_sd_categories}
\title{to_sd_categories}
\usage{
to_sd_categories(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector of numeric values}

\item{na.rm}{If true, remove NAs when calculating mean and SD. If NAs are present and na.rm is false, output will be all NAs}
}
\value{
Vector of 3 level factor by standard deviation.
}
\description{
Convert a numeric into a factor divided into three categories according to standard deviation (Below 1 SD, within 1 SD, greater than 1 SD).
Useful for plotting interactions with continuous variables.
}
\examples{
myVec <- c(1,6,3,6,8,4,19,5,4,1,7,17,20,9,5,5,11,4,3,4,11,4,2,2)
myCategoricalVec <- to_sd_categories(myVec)

}
