% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix_kbl.R
\name{cor_matrix_kbl}
\alias{cor_matrix_kbl}
\title{cor_matrix_kbl}
\usage{
cor_matrix_kbl(column_grid, dat, alpha = 0.05)
}
\arguments{
\item{column_grid}{dataframe: first two columns contain the names of pairwise variables found in dat}

\item{dat}{dataframe: data}

\item{alpha}{numeric: p-value cutoff at which to start coloring cells}
}
\value{

}
\description{
This function creates a kable correlation matrix with colored cells for
}
